Item 6944: Add ability to edit parent samples (#212)
