+++ packages/components/src/components/entities/EntityInsertPanel.tsx
@@ -154,7 +154,8 @@ export class EntityInsertPanelImpl extends React.Component<Props, StateProps> {
     }
 
     componentWillReceiveProps(nextProps: OwnProps) {
-        this.init(nextProps)
+        if (this.props.location != nextProps.location || this.props.entityDataType !== nextProps.entityDataType)
+            this.init(nextProps)
     }
 
     componentWillUnmount() {
