+++ packages/components/src/components/entities/actions.ts
@@ -1,18 +1,8 @@
-import {
-    buildURL,
-    getQueryGridModel,
-    getSelected,
-    naturalSort,
-    queryGridInvalidate,
-    QueryGridModel,
-    SchemaQuery,
-    selectRows
-} from '../..';
+import { buildURL, getQueryGridModel, getSelected, naturalSort, SchemaQuery, selectRows } from '../..';
 import { Ajax, Filter, Utils } from '@labkey/api';
 import { fromJS, List, Map } from 'immutable';
 import {
     DisplayObject,
-    EntityChoice,
     EntityDataType,
     EntityIdCreationModel,
     EntityParentType,
@@ -21,7 +11,6 @@ import {
     IParentOption
 } from './models';
 import { DataClassDataType, SampleTypeDataType } from './constants';
-import { DELIMITER } from '../forms/input/SelectInput';
 
 export interface DeleteConfirmationData {
     canDelete: Array<any>
@@ -290,78 +279,6 @@ export function getEntityTypeData(model: EntityIdCreationModel, entityDataTypes:
     });
 }
 
-export function getInitialParentChoices(parentTypeOptions: List<IEntityTypeOption>, parentDataType: EntityDataType, childModel: QueryGridModel) : List<EntityChoice> {
-    let parentValuesByType = Map<string, EntityChoice>();
-
-    if (childModel && childModel.isLoaded) {
-        const row = childModel.getRow();
-
-        if (row.size > 0) {
-            const inputs: List<Map<string, any>> = row.get(parentDataType.inputColumnName);
-            const inputTypes: List<Map<string, any>> = row.get(parentDataType.inputTypeColumnName);
-            if (inputs && inputTypes) {
-
-                // group the inputs by parent type so we can show each in its own grid.
-                inputTypes.forEach((typeMap, index) => {
-                    // I'm not sure when the type could have more than one value here, but 'value' is an array
-                    const typeValue = typeMap.getIn(['value', 0]);
-                    const typeOption = parentTypeOptions.find((option) => option[parentDataType.inputTypeValueField] === typeValue);
-                    if (!typeOption) {
-                        console.warn("Unable to find parent type.", typeValue);
-                    }
-                    else {
-                        if (!parentValuesByType.has(typeOption.query)) {
-                            parentValuesByType = parentValuesByType.set(typeOption.query, {
-                                type: typeOption,
-                                ids: [],
-                                value: undefined
-                            })
-                        }
-                        let updatedChoice = parentValuesByType.get(typeOption.query);
-                        updatedChoice.ids.push(inputs.getIn([index, 'value']));
-                        parentValuesByType = parentValuesByType.set(typeOption.query, updatedChoice)
-                    }
-                });
-            }
-        }
-    }
-    // having collected the values by type, create a list, sorted by the type label and return that.
-    return parentValuesByType.sortBy(choice => choice.type.label, naturalSort).toList();
-}
-
-export function getUpdatedRowForParentChanges(parentDataType: EntityDataType, originalParents: List<EntityChoice>, currentParents: List<EntityChoice>, childModel: QueryGridModel) {
-    const queryData = childModel.getRow();
-    const queryInfo = childModel.queryInfo;
-
-    const definedCurrentParents = currentParents.filter((parent) => (parent.type !== null && parent.type !== undefined)).toList();
-    let updatedValues = {};
-    if (definedCurrentParents.isEmpty()) { // have no current parents but have original parents, send in empty strings so original parents are removed.
-        originalParents.forEach((parentChoice) => {
-            updatedValues[parentDataType.insertColumnNamePrefix + parentChoice.type.label] = null;
-        })
-    }
-    else {
-        definedCurrentParents.forEach((parentChoice) => {
-            // Label may seem wrong here, but it is the same as query when extracted from the original query to get
-            // the entity types.
-            updatedValues[parentDataType.insertColumnNamePrefix + parentChoice.type.label] = parentChoice.value || null;
-        });
-    }
-
-    queryInfo.getPkCols().forEach((pkCol) => {
-        const pkVal = queryData.getIn([pkCol.fieldKey, 'value']);
-
-        if (pkVal !== undefined && pkVal !== null) {
-            updatedValues[pkCol.fieldKey] = pkVal;
-        }
-        else {
-            console.warn('Unable to find value for pkCol \"' + pkCol.fieldKey + '\"');
-        }
-    });
-    return updatedValues;
-
-}
-
 export function deleteEntityType(deleteActionName: string, rowId: number): Promise<any> {
     return new Promise((resolve, reject) => {
         return Ajax.request({
@@ -381,42 +298,3 @@ export function deleteEntityType(deleteActionName: string, rowId: number): Promi
     });
 }
 
-export function parentValuesDiffer(sortedOriginalParents: List<EntityChoice>, currentParents: List<EntityChoice>) : boolean {
-    const sortedCurrentParents = currentParents.sortBy(choice => choice.type ? choice.type.label : "~~NO_TYPE~~", naturalSort).toList();
-    const difference = sortedOriginalParents.find((original, index) => {
-        const current = sortedCurrentParents.get(index);
-        if (!current)
-            return true;
-        if (current.type && original.type.rowId !== current.type.rowId) {
-            return true;
-        }
-        const originalValues = original.value ? original.value.split(DELIMITER).sort(naturalSort).join(DELIMITER) : "";
-        const currentValues = current.value ? current.value.split(DELIMITER).sort(naturalSort).join(DELIMITER) : "";
-        if (originalValues !== currentValues) {
-            return true;
-        }
-    });
-    if (difference) {
-        return true;
-    }
-    // we have more current parents than the original and we have selected a value for at least one of these parents.
-    if (sortedCurrentParents.size > sortedOriginalParents.size) {
-        return sortedCurrentParents.slice(sortedOriginalParents.size).find((parent) => parent.value !== undefined) !== undefined;
-    }
-    return false;
-}
-
-export function invalidateParentModels(originalParents: List<EntityChoice>, currentParents: List<EntityChoice>, parentDataType: EntityDataType) {
-    // clear out the original parents' grid data (which may no longer be represented in the current parents)
-    let cleared = [];
-    originalParents.forEach((parentChoice) => {
-        cleared.push(parentChoice.type.label);
-        queryGridInvalidate(SchemaQuery.create(parentDataType.instanceSchemaName, parentChoice.type.label), true);
-    });
-    // also clear out the current parents' grid data if it hasn't already been cleared
-    currentParents.forEach((parentChoice) => {
-        if (parentChoice.type && cleared.indexOf(parentChoice.type.label) < 0) {
-            queryGridInvalidate(SchemaQuery.create(parentDataType.instanceSchemaName, parentChoice.type.label), true);
-        }
-    });
-}
