+++ packages/components/src/components/entities/constants.ts
@@ -4,6 +4,8 @@ import { EntityDataType } from './models';
 export const DATA_DELETE_CONFIRMATION_ACTION = 'getDataDeleteConfirmationData.api';
 export const SAMPLE_DELETE_CONFIRMATION_ACTION = 'getMaterialDeleteConfirmationData.api';
 
+export const PARENT_DATA_GRID_PREFIX = 'parent-data';
+
 export const SampleTypeDataType : EntityDataType = {
     typeListingSchemaQuery: SCHEMAS.EXP_TABLES.SAMPLE_SETS,
     instanceSchemaName: SCHEMAS.SAMPLE_SETS.SCHEMA,
@@ -19,7 +21,7 @@ export const SampleTypeDataType : EntityDataType = {
     deleteHelpLinkTopic: DELETE_SAMPLES_TOPIC,
     inputColumnName: "Inputs/Materials/First",
     inputTypeColumnName: "Inputs/Materials/First/SampleSet",
-    inputTypeValueField: "LSID",
+    inputTypeValueField: "lsid",
     insertColumnNamePrefix: "MaterialInputs/",
 };
